kmsdrm: Finetune integer type usage. Add some comments.

+++ src/video/kmsdrm/SDL_kmsdrmmouse.h
@@ -35,8 +35,8 @@ typedef struct _KMSDRM_CursorData
     struct gbm_bo *bo;
     struct plane *plane;
     uint32_t       crtc_id;
-    int            hot_x, hot_y;
-    int            w, h;
+    uint16_t       hot_x, hot_y;
+    uint16_t       w, h;
     /* The video devide implemented on SDL_kmsdrmvideo.c 
      * to be used as _THIS pointer in SDL_kmsdrmvideo.c 
      * functions that need it. */
